+++ programs/dynamic-fee-sharing/src/error.rs
@@ -43,4 +43,7 @@ pub enum FeeVaultError {
 
     #[msg("Type cast error")]
     TypeCastFailed,
+
+    #[msg("Fee vault is not mutable")]
+    FeeVaultNotMutable,
 }