+++ programs/dynamic-fee-sharing/src/state/fee_vault.rs
@@ -33,7 +33,8 @@ pub struct FeeVault {
     pub token_flag: u8, // indicate whether token is spl-token or token2022
     pub fee_vault_type: u8,
     pub fee_vault_bump: u8,
-    pub padding_0: [u8; 13],
+    pub mutable_flag: u8, // indicate whether the fee vault is mutable by admin or operator, 0 or 1 only
+    pub padding_0: [u8; 12],
     pub total_share: u32,
     pub padding_1: [u8; 4],
     pub total_funded_fee: u64,
@@ -90,6 +91,7 @@ impl FeeVault {
         fee_vault_bump: u8,
         fee_vault_type: u8,
         users: &[UserShare],
+        mutable_flag: u8,
     ) -> Result<()> {
         self.owner = *owner;
         self.token_flag = token_flag;
@@ -109,6 +111,7 @@ impl FeeVault {
         self.fee_vault_bump = fee_vault_bump;
         self.fee_vault_type = fee_vault_type;
         self.operator = Pubkey::default();
+        self.mutable_flag = mutable_flag;
 
         Ok(())
     }