+++ tests/fee_sharing.test.ts
@@ -74,6 +74,7 @@ describe("Fee vault sharing", () => {
 
     const params: InitializeFeeVaultParameters = {
       padding: [],
+      mutableFlag: false,
       users,
     };
 
@@ -106,6 +107,7 @@ describe("Fee vault sharing", () => {
 
     const params: InitializeFeeVaultParameters = {
       padding: [],
+      mutableFlag: false,
       users,
     };
 
@@ -142,6 +144,7 @@ describe("Fee vault sharing", () => {
 
     const params: InitializeFeeVaultParameters = {
       padding: [],
+      mutableFlag: false,
       users,
     };
 
@@ -169,6 +172,57 @@ describe("Fee vault sharing", () => {
     expectThrowsErrorCode(svm.sendTransaction(tx), errorCode);
   });
 
+  it("Fail to update operator when fee vault is not mutable", async () => {
+    const generatedUser = generateUsers(svm, 5);
+    const users = generatedUser.map((item) => ({
+      address: item.publicKey,
+      share: 1000,
+    }));
+
+    const params: InitializeFeeVaultParameters = {
+      mutableFlag: false,
+      padding: [],
+      users,
+    };
+
+    const feeVault = Keypair.generate();
+    const tokenVault = deriveTokenVaultAddress(feeVault.publicKey);
+    const feeVaultAuthority = deriveFeeVaultAuthorityAddress();
+
+    const tx = await program.methods
+      .initializeFeeVault(params)
+      .accountsPartial({
+        feeVault: feeVault.publicKey,
+        feeVaultAuthority,
+        tokenVault,
+        tokenMint,
+        owner: vaultOwner.publicKey,
+        payer: admin.publicKey,
+        tokenProgram: TOKEN_PROGRAM_ID,
+      })
+      .transaction();
+
+    tx.recentBlockhash = svm.latestBlockhash();
+    tx.sign(admin, feeVault);
+    const initializeFeeVaultRes = svm.sendTransaction(tx);
+    expect(initializeFeeVaultRes instanceof TransactionMetadata).to.be.true;
+
+    const errorCode = getProgramErrorCodeHexString("FeeVaultNotMutable");
+
+    const updateOperatorTx = await program.methods
+      .updateOperator()
+      .accountsPartial({
+        feeVault: feeVault.publicKey,
+        operator: user.publicKey,
+        owner: vaultOwner.publicKey,
+      })
+      .transaction();
+    updateOperatorTx.recentBlockhash = svm.latestBlockhash();
+    updateOperatorTx.sign(vaultOwner);
+    const updateOperatorRes = svm.sendTransaction(updateOperatorTx);
+    expectThrowsErrorCode(updateOperatorRes, errorCode);
+  });
+
   it("Fail to perform admin task when not an admin", async () => {
     const generatedUser = generateUsers(svm, 5);
     const users = generatedUser.map((item) => ({
@@ -178,6 +232,7 @@ describe("Fee vault sharing", () => {
 
     const params: InitializeFeeVaultParameters = {
       padding: [],
+      mutableFlag: true,
       users,
     };
 
@@ -262,6 +317,7 @@ describe("Fee vault sharing", () => {
 
     const params: InitializeFeeVaultParameters = {
       padding: [],
+      mutableFlag: true,
       users,
     };
 
@@ -322,6 +378,7 @@ describe("Fee vault sharing", () => {
 
     const params: InitializeFeeVaultParameters = {
       padding: [],
+      mutableFlag: true,
       users,
     };