+++ tests/fee_sharing_pda.test.ts
@@ -77,6 +77,7 @@ describe("Fee vault pda sharing", () => {
 
     const params: InitializeFeeVaultParameters = {
       padding: [],
+      mutableFlag: false,
       users,
     };
 
@@ -110,6 +111,7 @@ describe("Fee vault pda sharing", () => {
 
     const params: InitializeFeeVaultParameters = {
       padding: [],
+      mutableFlag: false,
       users,
     };
 
@@ -147,6 +149,7 @@ describe("Fee vault pda sharing", () => {
 
     const params: InitializeFeeVaultParameters = {
       padding: [],
+      mutableFlag: false,
       users,
     };
 
@@ -175,6 +178,58 @@ describe("Fee vault pda sharing", () => {
     expectThrowsErrorCode(svm.sendTransaction(tx), errorCode);
   });
 
+  it("Fail to update operator when fee vault is not mutable", async () => {
+    const generatedUser = generateUsers(svm, 5);
+    const users = generatedUser.map((item) => ({
+      address: item.publicKey,
+      share: 1000,
+    }));
+
+    const params: InitializeFeeVaultParameters = {
+      mutableFlag: false,
+      padding: [],
+      users,
+    };
+
+    const feeVault = deriveFeeVaultPdaAddress(baseKp.publicKey, tokenMint);
+    const tokenVault = deriveTokenVaultAddress(feeVault);
+    const feeVaultAuthority = deriveFeeVaultAuthorityAddress();
+
+    const tx = await program.methods
+      .initializeFeeVaultPda(params)
+      .accountsPartial({
+        feeVault,
+        base: baseKp.publicKey,
+        feeVaultAuthority,
+        tokenVault,
+        tokenMint,
+        owner: vaultOwner.publicKey,
+        payer: admin.publicKey,
+        tokenProgram: TOKEN_PROGRAM_ID,
+      })
+      .transaction();
+
+    tx.recentBlockhash = svm.latestBlockhash();
+    tx.sign(admin, baseKp);
+    const initializeFeeVaultRes = svm.sendTransaction(tx);
+    expect(initializeFeeVaultRes instanceof TransactionMetadata).to.be.true;
+
+    const errorCode = getProgramErrorCodeHexString("FeeVaultNotMutable");
+
+    const updateOperatorTx = await program.methods
+      .updateOperator()
+      .accountsPartial({
+        feeVault,
+        operator: user.publicKey,
+        owner: vaultOwner.publicKey,
+      })
+      .transaction();
+    updateOperatorTx.recentBlockhash = svm.latestBlockhash();
+    updateOperatorTx.sign(vaultOwner);
+    const updateOperatorRes = svm.sendTransaction(updateOperatorTx);
+    expectThrowsErrorCode(updateOperatorRes, errorCode);
+  });
+
   it("Fail to perform admin task when not an admin", async () => {
     const generatedUser = generateUsers(svm, 5);
     const users = generatedUser.map((item) => ({
@@ -184,6 +239,7 @@ describe("Fee vault pda sharing", () => {
 
     const params: InitializeFeeVaultParameters = {
       padding: [],
+      mutableFlag: true,
       users,
     };
 
@@ -269,6 +325,7 @@ describe("Fee vault pda sharing", () => {
 
     const params: InitializeFeeVaultParameters = {
       padding: [],
+      mutableFlag: true,
       users,
     };
 
@@ -330,6 +387,7 @@ describe("Fee vault pda sharing", () => {
 
     const params: InitializeFeeVaultParameters = {
       padding: [],
+      mutableFlag: true,
       users,
     };